fix: upgrade to latest jx
and minor refactor of code
